New model for database

+++ src/main/java/com/assist/bookingjava/DataBase/CustomerDao.java
@@ -9,5 +9,5 @@ import java.util.List;
  * Created by doroftei on 13.07.2017.
  */
 public interface CustomerDao extends CrudRepository<Customer,Long>{
-    List<Customer>findById(Integer id);
+    List<Customer>findById(Long id);
 }
